Fix typespecs on `change_user/0` - `change_user/2`
Per #20. I experimented with using `%@user_schema{}` for the typespecs,
but got weird errors, so I decided to keep the `.t` references for now.

I think these tweaks will fix the problem Dialyzer is reporting in #20.

I added `@type t :: %__MODULE__{}` lines to all the example schemas in
the documentation as well.